+++ index.html
@@ -6,7 +6,8 @@
         <link rel="stylesheet" href="base.css">
     </head>
     <body>
-        <script src="./main.js" type="module"></script>
+        <script src="js/main.js" type="module"></script>
+
     </body>
 
 </html>
